API tweaks and module tidy up
