+++ treetop_test.go
@@ -5,10 +5,10 @@ import (
 	"testing"
 )
 
-func Test_renderer_Define(t *testing.T) {
-	renderer := NewRenderer(DefaultTemplateExec)
+func Test_treetop_Define(t *testing.T) {
+	tt := NewTreetop(DefaultTemplateExec)
 
-	page := renderer.Define("base.templ.html", Noop)
+	page := tt.Define("base.templ.html", Noop)
 
 	a := page.Block("A")
 	b := page.Block("B")
@@ -20,15 +20,15 @@ func Test_renderer_Define(t *testing.T) {
 
 	expects := []string{"b.templ.html"}
 	if files, err := handler.Partial.TemplateList(); err != nil {
-		t.Errorf("renderer.Define() Partial = unexpected error %s", err.Error())
+		t.Errorf("treetop.Define() Partial = unexpected error %s", err.Error())
 	} else if !reflect.DeepEqual(files, expects) {
-		t.Errorf("renderer.Define() = %v, want %v", files, expects)
+		t.Errorf("treetop.Define() = %v, want %v", files, expects)
 	}
 
 	expects = []string{"base.templ.html", "a.templ.html", "b.templ.html"}
 	if files, err := handler.Page.TemplateList(); err != nil {
-		t.Errorf("renderer.Define() Page = unexpected error %s", err.Error())
+		t.Errorf("treetop.Define() Page = unexpected error %s", err.Error())
 	} else if !reflect.DeepEqual(files, expects) {
-		t.Errorf("renderer.Define() = %v, want %v", files, expects)
+		t.Errorf("treetop.Define() = %v, want %v", files, expects)
 	}
 }
